% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_overlaps.R
\name{find_overlaps}
\alias{find_overlaps}
\title{Find overlaps}
\usage{
find_overlaps(data)
}
\arguments{
\item{data}{A tibble of annotations}
}
\value{
A tibble with added columns of overlaps (self and other)
}
\description{
Find annotations that overlap, whether within or across participants.
Inputs a data frame structured like those returned from `read_eaf()`
}
