% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_turns.R
\name{find_turns}
\alias{find_turns}
\title{Find turns}
\usage{
find_turns(data, method = "sequential", interval = 500, simplify = FALSE)
}
\arguments{
\item{data}{A tibble of annotations}

\item{method}{Method for inferring turns, "interval" or "sequential" (default = "sequential")}

\item{interval}{Threshold for inferring turns with "interval" method (default = 500)}

\item{simplify}{Whether to collapse rows to turns only (default = `FALSE`)}
}
\value{
A tibble with inferred turns added
}
\description{
Infer turns from annotations in data frames like those returned by `read_eaf()`.
Uses one of two methods ("interval" or "sequential") to infer turns.
}
