% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cmdi.R
\name{read_cmdi}
\alias{read_cmdi}
\title{Read CMDI files}
\usage{
read_cmdi(path, meta = "actor", recursive = TRUE)
}
\arguments{
\item{path}{Path to CMDI (.cmdi) metadata file(s)}

\item{meta}{Which metadata to be collected: "actor"/"content" (default = `"actor"`)}

\item{recursive}{CMDI (.cmdi) files are read recursively in path (defaults to `TRUE`)}
}
\value{
Tibble of the CMDI data
}
\description{
Reads CMDI (.cmdi) file(s) from a path input.
Either "actor" or "content" metadata can be
read, based on the `meta` argument input.
}
