% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eaf.R
\name{read_eaf}
\alias{read_eaf}
\title{Read EAF files}
\usage{
read_eaf(path, tiers = c(), xpath = "", recursive = TRUE)
}
\arguments{
\item{path}{Path to ELAN (.eaf) file(s)}

\item{tiers}{Specify tiers (TIER_ID or TIER_TYPE) to be read (default = `c()`)}

\item{xpath}{Specify a detailed XPath for tiers to be read}

\item{recursive}{ELAN (.eaf) files are read recursively in path (defaults to `TRUE`)}
}
\value{
Tibble of the ELAN annotations
}
\description{
Reads ELAN (.eaf) annotation file(s) from a path input,
whether a direct filename or a directory of multiple files.
Specific tiers or tier types can be targeted through:
`tiers = c("tier_type" = "your_tiername", "tier_type" = "your_tier_type")`
}
