% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ilex.R
\name{read_ilex}
\alias{read_ilex}
\title{Read iLex files}
\usage{
read_ilex(path, recursive = TRUE)
}
\arguments{
\item{path}{Path to iLex (.ilex) file(s)}

\item{recursive}{iLex (.ilex) files are read recursively in path (defaults to `TRUE`)}
}
\value{
Tibble of the iLex annotations
}
\description{
[Experimental]
Reads iLex (.ilex) file(s) from a path input
Note: Not extensively tested!
}
