% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_overlaps.R
\name{find_overlaps}
\alias{find_overlaps}
\title{Find overlaps in annotations within and across participants}
\usage{
find_overlaps(data)
}
\arguments{
\item{data}{A tibble of annotations}
}
\value{
A tibble with added columns of overlaps (self and other)
}
\description{
Find overlaps in annotations within and across participants
}
