% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_children.R
\name{pivot_children}
\alias{pivot_children}
\title{Pivot child tier annotations}
\usage{
pivot_children(data)
}
\arguments{
\item{data}{A tibble of annotations}
}
\value{
A tibble with dependent (child) annotations in wide format
}
\description{
[Experimental]
Pivot ELAN annotations from child tiers to a wide format.
Note: Pivots empty values as NA! A better option may be to use
`dplyr::_join()` to join relevant child tier annotations to their
associated parents, or to `dplyr::rename()` columns before pivoting.
}
