% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cmdi.R
\name{read_cmdi}
\alias{read_cmdi}
\title{Reads CMDI (.cmdi) file(s) from a path input}
\usage{
read_cmdi(path, meta = "signers", recursive = TRUE)
}
\arguments{
\item{path}{Path to CMDI (.cmdi) metadata file(s)}

\item{meta}{Which metadata to be collected: "signers"/"content" (default = `"signers"`)}

\item{recursive}{CMDI (.cmdi) files are read recursively in path (defaults to `TRUE`)}
}
\value{
Tibble of the CMDI data
}
\description{
Reads CMDI (.cmdi) file(s) from a path input
}
