% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eaf.R
\name{read_eaf}
\alias{read_eaf}
\title{Reads ELAN (.eaf) file(s) from a path input}
\usage{
read_eaf(path, tiers = c(), xpath = "", recursive = TRUE)
}
\arguments{
\item{path}{Path to ELAN (.eaf) file(s)}

\item{tiers}{Specify tiers (TIER_ID or TIER_TYPE) to be read (default = `c()`)}

\item{xpath}{Specify a detailed XPath for tiers to be read}

\item{recursive}{ELAN (.eaf) files are read recursively in path (defaults to `TRUE`)}
}
\value{
Tibble of the ELAN annotations
}
\description{
Reads ELAN (.eaf) file(s) from a path input
}
